apply lint on project

+++ app/src/main/java/com/droids/tamada/filemanager/helper/Utilities.java
@@ -62,7 +62,7 @@ public class Utilities {
      */
     public int progressToTimer(int progress, int totalDuration) {
         int currentDuration = 0;
-        totalDuration = (int) (totalDuration / 1000);
+        totalDuration = totalDuration / 1000;
         currentDuration = (int) ((((double) progress) / 100) * totalDuration);
 
         // return current duration in milliseconds
